move ProgramViewSet stub to v1/views/__init__.py; empty programs/__init__.py
